+++ test/errors.js
@@ -1,0 +1,7 @@
+const { join } = require('path');
+const { errors } = require('../lib');
+const { CodedError } = errors;
+
+errors.save(new Error("Something went wrong in X."), 'X-%s', "An example.", join(__dirname, 'errors'));
+
+console.error(new CodedError('some-code', 'MASSIVE ERROR!'));