+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 Coming soon...
 
+## [1.2.1] - 2023-06-00
+
+### Changed
+- Save Error Module to be Errors Module.
+
+### Added
+- Coded Error.
+
 ## [1.2.0]
 
 ### Added